Adding basic field validations for events
I removed Dan's module which was WAY more work than was needed.
I created some basic validations for title, end date and description. I
also modified some of the fields. Each field should probably have some
help text and target product should probably be a list of products, or
maybe a reference to products after those are in Drupal (I think that
would be really cool as we could then find events that reference a
product and display that to the users we view a product).